+++ src/commands/guess.ts
