+++ src/commands/guess/special-handler.ts
