+++ src/commands/guess/update-stats.ts
