+++ src/commands/numberdex.ts
