+++ src/commands/numberdex/numberhumans.ts
