+++ src/commands/numberdex/show-humans.ts
