+++ src/commands/poweroff.ts
