+++ src/commands/users/show.ts
