+++ src/index.ts
@@ -6,7 +6,7 @@
  */
 import { Logger } from "#utils/logger.ts";
 import type { Client } from "discord.js";
-import { registerCommands } from "./commands/listener.ts";
+import { registerCommands } from "./cmd-handler/listener.ts";
 import { registerHandlers } from "./handlers.ts";
 
 export async function initClient(client: Client, token: string): Promise<void> {
