+++ src/main.ts
@@ -13,8 +13,8 @@ import type { Command as ApplicationCommand } from "#utils/types.ts";
 import { Command } from "commander";
 import { Client } from "discord.js";
 import packageJson from "../package.json" with { type: "json" };
-import type { GuessCooldownCollection } from "./commands/cooldowns/guesses.ts";
-import type { CooldownCollection } from "./commands/cooldowns/normal.ts";
+import type { GuessCooldownCollection } from "./cmd-handler/cooldowns/guesses.ts";
+import type { CooldownCollection } from "./cmd-handler/cooldowns/normal.ts";
 import { initClient } from "./index.ts";
 
 const program = new Command()
