+++ tsconfig.json
@@ -36,7 +36,7 @@
       "#utils/*": ["./src/utils/*"],
       "#numberdex/*": ["./src/numberdex/*"],
       "#fg-sparky/*": ["./src/sparky/*"],
-      "#cmd-loader": ["./src/commands/loader.ts"]
+      "#cmd-loader": ["./src/cmd-handler/loader.ts"]
     }
   },
   "include": ["**/*.ts"]
